fix: resolve 10 golangci-lint issues from feat/clickhouse-sink
- copyloopvar: remove redundant `i := i` in buffer_test.go (Go 1.22+)
- errcheck: use `_, _ =` to explicitly discard fmt.Fprintf/Fprintln
  return values in schema_cli.go
- staticcheck QF1006: lift ctx.Err()/BufferLen() checks into for-loop
  conditions in writer.go and writer_test.go
- unconvert: remove unnecessary int32() cast in tables_test.go

+++ internal/sink/clickhouse/tables_test.go
@@ -585,7 +585,7 @@ func makeSession(id int32) Session {
 	return Session{
 		SnapshotTime: now,
 		SessionID:    id,
-		PID:          int32(10000 + id),
+		PID:          10000 + id,
 		User:         "alice",
 		Database:     "warehouse",
 		Application:  "psql",
